Django/app/requirements_django.txt
alabaster==0.7.16
altair==5.0.1
anyio==4.4.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.2.1
attrs==23.2.0
automium==0.2.6
automium-web==0.1.1
Babel==2.15.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bibtexparser==1.4.1
biopython==1.81
blinker==1.6.2
bokeh==3.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
chatbot==1.5.2b0
click==8.1.3
cloudbiolinux==0.1
comm==0.1.3
config==0.5.1
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cycler==0.11.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
Django==3.2.18
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-extensions==3.2.3
django-pgconnection==1.0.3
django-plotly-dash==2.1.4
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
dpd-components==0.1.0
drf-yasg==1.21.7
drf-yasg2==1.19.4
ete3==3.1.3
exceptiongroup==1.2.1
executing==1.2.0
fabric==3.0.0
fake-useragent==1.5.1
feedparser==6.0.11
Flask==2.2.3
fonttools==4.39.3
free-proxy==1.1.1
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
inflection==0.5.1
install==1.3.5
invoke==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
legacy==0.1.6
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
networkx==3.2.1
numpy==1.24.2
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==23.0
pandas==2.0.0
paramiko==3.1.0
parso==0.8.3
pexpect==4.8.0
pgconn==0.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.14.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
pydotplus==2.0.2
Pygments==2.15.0
PyJWT==2.8.0
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
referencing==0.30.2
requests==2.30.0
rich==13.5.2
rpds-py==0.9.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scholarly==1.7.11
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
selenium==4.21.0
sgmllib3k==1.0.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.3
stack-data==0.6.2
streamlit==1.40.1
tenacity==8.2.2
threadpoolctl==3.2.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.0.2
validators==0.21.2
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
whitenoise==6.4.0
widgetsnbextension==4.0.7
wrapt==1.12.1
wsproto==1.2.0
xyzservices==2023.2.0
zipp==3.16.2
